requirements.txt
anndata
autopep8
bcftools
black
bwa
fa2
fa2_modified
flake8
ipympl
ipython
jedi
jupyter-lsp
jupyterlab
jupyterlab_latex
jupyterlab_vim
jupyterlab_widgets
leidenalg
loky
loompy
louvain
matplotlib
minimap2
mygene
numpy
pandas
pep8
polars
pyarrow
pycairo
pycodestyle
pyflakes
pylint
pynvim
pyqt
python
python-graphviz
qt-main
samtools
scanpy
scikit-image
scikit-learn
scikit-misc
scipy
scvelo
seaborn
spacy
statsmodels
tabix
toolz
tsne
umap-learn
xlsx2csv
yapf
